requirements.txt
flask
twilio
sentence-transformers
faiss-cpu
pandas
numpy
openai
prophet
datetime
json
typing
os
sys
pickle
math
ibm_granite
re
ibm-watsonx-ai